Fix support for writing more than 65535 items
A number of boxes wrote the itemId values as an UInt16, these boxes
will now use an UInt32 itemId if it is required.

This change also includes a performance optimization that preallocates
the ItemLocationBox and ItemInfoBox lists with the total number of items.
Every item that is saved will have ItemLocationBox and ItemInfoBox
entries.

+++ src/Avif Container/Boxes/Item Info Entries/AV01ItemInfoEntryBox.cs
@@ -15,7 +15,7 @@ namespace AvifFileType.AvifContainer
     internal sealed class AV01ItemInfoEntryBox
         : ItemInfoEntryBox
     {
-        public AV01ItemInfoEntryBox(ushort itemId, string name)
+        public AV01ItemInfoEntryBox(uint itemId, string name)
             : base(itemId, 0, ItemInfoEntryTypes.AV01, name)
         {
         }
